requirements.txt
# Core ML
scikit-learn>=1.8.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Deep Learning (optional DistilBERT path)
torch>=2.0.0
transformers>=4.35.0

# HuggingFace datasets
datasets>=2.14.0

# MLflow
mlflow>=2.8.0
boto3>=1.28.0            # for S3 artifact storage (optional)

# Evidently AI
evidently>=0.4.11

# FastAPI serving
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# DVC
dvc>=3.25.0
dvc-s3>=2.23.0           # optional: for remote DVC storage

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0            # async test client for FastAPI

# Utilities
pyyaml>=6.0
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-multipart>=0.0.6

# Notebook (optional)
jupyter>=1.0.0
ipykernel>=6.25.0
